hw updates
Project proposal is uploaded.

I attempted to make some changes in the HW due to your feedback but I may have made it worse.  I keep on getting errors in the console however I tried to place the whole formula within the containers.

I know we dont need to use concantation, however if I already have it labeled there and its working would it be ok to leave it?

looking forward to the soultion review this Monday.

Best,
Brittany